cambios en cardDetalle, CardSinopsis y App.js
