--- src/Components/Card/CardDetail/CardDetail.css
@@ -1,61 +1,0 @@
-/* @import url("https://fonts.googleapis.com/css?family=Roboto+Slab:100,300,400,700");
-@import url("https://fonts.googleapis.com/css?family=Raleway:300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i"); */
-
-.card-movie {
-  display: flex;
-  width: 100%;
-  margin: 1rem;
-  padding: 1rem;
-  border: 4px solid var(--color-naranja);
-  border-radius: 2vh;
-  background-color: var(--color-azul);
-}
-
-.img-container {
-  width: 50%;
-}
-
-.img-pelicula {
-  width: 100%;
-  padding: 0.1vh;
-  object-fit: cover;
-  border-radius: 10px;
-}
-
-.info-container {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-}
-
-.btn-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 1rem;
-}
-
-.titulo {
-  font-size: 30px;
-  text-align: center;
-  font-weight: bold;
-}
-
-.datos-container {
-  display: flex;
-  text-align: center;
-  flex-direction: column;
-  justify-content: center;
-  color: white;
-  padding: 1rem;
-}
-
-.info-movie {
-  display: flex;
-  font-size: 25px;
-  flex-direction: column;
-}
-
-p:hover {
-  color: white;
-}
