+++ src/pages/Card/CardSinopsis.jsx
@@ -1,30 +1,40 @@
-import React from 'react'
+import { useState, useEffect } from "react";
+import { useParams } from "react-router-dom";
+import { getMovieById, getTVById } from "../../services/tmdbService";
 import LayoutForm from '../../layout/LayoutForm/LayoutForm'
-import CardDetalle from  '../../components/Card/CardDetalle/CardDetalle'
-import { useLocation } from 'react-router-dom';
+import CardDetalle from '../../components/Card/CardDetalle/CardDetalle'
 
-const CardSinopsis = () => { 
-    const { state } = useLocation();
-    const dato = state && state.dato;
-    const selectedMovie = state && state.dato;
+const CardSinopsis = () => {
+  //   Leo el parametro id
+  const { id } = useParams();
+  const [info, setInfo] = useState(null);
+  const isMovie = window.location.pathname.includes("/movie/"); // Corregido
 
-    if (!dato) {
-        // Manejo si el objeto dato no está presente
-        return <div>No hay datos disponibles.</div>;
+  // Función para buscar la película o serie
+  const fetchData = async () => {
+    try {
+      const result = isMovie ? await getMovieById(id) : await getTVById(id);
+      setInfo(result);
+    } catch (error) {
+      console.error(
+        `Error buscando ${
+          isMovie ? "la película" : "la serie de televisión"
+        } (CardSinopsis2)`,
+        error
+      );
     }
+  };
 
-    if (!selectedMovie) {
-        // Manejo si el objeto dato no está presente
-        return <div>No hay datos disponibles.</div>;
-    }
+  useEffect(() => {
+    fetchData();
+  }, [id]);
 
-    return (
-        <LayoutForm>
-            <CardDetalle info={dato} />
-            {/* {<Banner/>} Backdrop de la card en width 769px
-            <Card/> poster titulo + subtitulo: genero + boton + sinopsis */}
-        </LayoutForm>
-    )
-}
+  return (
+    <LayoutForm>
+      {/* Renderizo el componente solo si el estado es distinto de null */}
+      {info !== null ? <CardDetalle movie={info} /> : null}
+    </LayoutForm>
+  );
+};
 
-export default CardSinopsis
+export default CardSinopsis;
